Move planetary features into their own directory
