+++ README.md
@@ -1,1 +1,16 @@
-# vast-and-starlit
+# Vast and Starlit
+
+## Setup
+
+```bash
+pip3 install -r requirements.txt
+```
+
+## Running the application
+
+
+## Create a test Vista
+
+```bash
+python3 -c "import space_vista as sv; sv.Vista(sv.Coordinates(), sv.ShipLocation()).save_video()"
+```
