+++ requirements.txt
@@ -3,8 +3,8 @@ Cython==0.29.21
 imageio==2.9.0
 imageio-ffmpeg==0.4.2
 moviepy==1.0.3
-numpy==1.19.2
-pandas==1.1.2
+numpy>=1.22.4
+pandas==1.4.2
 Pillow==8.0.1
 pydub==0.24.1
 requests==2.24.0
