+++ space_vista.py
@@ -21,6 +21,11 @@ from PIL import (
     ImageFont,
 )
 from tqdm import tqdm
+from planetary_features.clouds import Clouds
+from planetary_features.continents import Continents
+from planetary_features.craters import Craters
+from planetary_features.rings import Rings
+from planetary_features.satellite_cluster import SatelliteCluster
 
 # Constants
 
@@ -631,612 +636,6 @@ class RandomPlanet(BasePlanet):
             features.append(SatelliteCluster(self))
         return features
 
-
-class PlanetaryFeature:
-    """
-    A basic class to subclass other planetary features from.
-    """
-
-    def __init__(self, planet):
-        self.planet = planet
-        self.sys_width = 0
-        self.sys_height = 0
-
-    def background(self, center, im, draw_im):
-        """
-        Called once to draw a feature that might be occluded by the planet on the
-        planet's base image.
-        Only used by features that do not move with respect to the planet.
-        """
-        pass
-
-    def foreground(self, center, im, draw_im, planet_mask):
-        """
-        Called once to draw a feature that will not be occluded by the planet on the
-        planet's base image.
-        Only used by features that do not move with respect to the planet.
-        """
-        pass
-
-    def drifting_background(self, x, y, image, drawing_frame, frame_n):
-        """
-        Called every frame to draw a feature that might be occuluded by the planet
-        as it moves with respect to that planet.
-        """
-        pass
-
-    def drifting_foreground(self, x, y, image, drawing_frame, frame_n):
-        """
-        Called every frame to draw a feature that will not be occuluded by the planet
-        as it moves with respect to that planet.
-        """
-        pass
-
-
-class SatelliteCluster(PlanetaryFeature):
-    """
-    A container for a planet's satellites that keeps them in line so that they overlap
-    each other in the proper order and maintain enough of a distance to avoid other
-    odd behaviors.
-    """
-
-    def __init__(self, planet):
-        super().__init__(planet)
-        self.band = set(range(self.planet.vista.length))
-        self.satellites = self.create_sats()
-
-    def create_sats(self):
-        """
-        Randomly determines the number of satellites orbiting a planet and assembles
-        a list of those satellites.
-
-        Returns a list of all the satellites
-        """
-        satellites = []
-        n_satellites = self.planet.vista.coords.choice(
-            [1, 2, 3, self.planet.vista.coords.integers(3, 12)],
-            p=[0.625, 0.125, 0.125, 0.125],
-        )
-        for moon in range(n_satellites):
-            satellites.append(Satellite(self.planet, self))
-        return satellites
-
-    def get_start_location(self, sat):
-        """
-        Called by the Satellite to find where in its orbit around the planet it will
-        start when we begin the animation. This method sets all the satellites in the
-        cluster with start_locations that ensure they will not overlap when directly
-        in front of the planet. This is necessary because this is the spot where two
-        satellites might appear to move through each other when one moves ahead of
-        the other in the sorting.
-        """
-        start_location = self.planet.vista.coords.choice(list(self.band))
-        while set(range(start_location, start_location + sat.mass)) - self.band:
-            start_location = self.planet.vista.coords.choice(list(self.band))
-        self.band -= set(range(start_location, start_location + sat.mass))
-        return start_location
-
-    def drifting_background(self, x, y, image, drawing_frame, frame_n):
-        """
-        Sorts the satellites in the background so that those closer to the viewer are
-        drawn last. Then calls on the Satellite's drifting_background method to draw it.
-        """
-        # The greater abs(cos) should be in front
-        def sort_key(sat):
-            phase = (
-                (sat.orbit_velocity * frame_n + sat.start_location)
-                / sat.planet.vista.length
-                * 2
-                * np.pi
-            )
-            return abs(np.cos(phase))
-
-        for sat in sorted(self.satellites, key=sort_key):
-            sat.drifting_background(x, y, image, drawing_frame, frame_n)
-
-    def drifting_foreground(self, x, y, image, drawing_frame, frame_n):
-        """
-        Sorts the satellites in the foreground so that those closer to the viewer are
-        drawn last. Then calls on the Satellite's drifting_foreground method to draw it.
-        """
-        # The lesser abs(cos) should be in front
-        def sort_key(sat):
-            phase = (
-                (sat.orbit_velocity * frame_n + sat.start_location)
-                / sat.planet.vista.length
-                * 2
-                * np.pi
-            )
-            return -abs(np.cos(phase))
-
-        for sat in sorted(self.satellites, key=sort_key):
-            sat.drifting_foreground(x, y, image, drawing_frame, frame_n)
-
-
-class Satellite(PlanetaryFeature):
-    """
-    The base class for a satellite. Right now, it's very much only moons. I'll have
-    to do some reworking of the code when I want to create artificial satellites.
-    """
-
-    def __init__(self, planet, cluster):
-        super().__init__(planet)
-        self.vista = self.planet.vista
-        self.colorful = self.planet.colorful
-        self.cluster = cluster
-
-        # It is necessary to set this to False, otherwise certain features will be
-        # reported as present in the Vista.status_dict.
-        self.is_planetoid = False
-
-        # Setting the appearance of the satellite
-        self.mass = max(
-            1,
-            self.planet.vista.coords.integers(
-                self.planet.mass // 10, self.planet.mass // 4
-            ),
-        )
-        if self.colorful:
-            self.hue = self.planet.vista.coords.choice(
-                [
-                    hue
-                    for hue in range(self.planet.vista.palette.hue_depth)
-                    if hue != self.planet.hue
-                ]
-            )
-            self.shade = self.planet.shade
-        else:
-            self.hue = self.planet.hue
-            self.shade = self.planet.vista.coords.choice(
-                [
-                    shade
-                    for shade in range(self.planet.vista.palette.shade_depth - 1)
-                    if shade != self.planet.shade
-                ]
-            )
-        self.fill = self.planet.vista.palette.get_color(hue=self.hue, shade=self.shade)
-        self.irregular = self.planet.vista.coords.choice([True, False], p=[0.1, 0.9])
-        if self.irregular:
-            self.vista.status_dict[f"Planet {self.planet.layer}"]["irregular moon"] = +1
-        else:
-            self.vista.status_dict[f"Planet {self.planet.layer}"]["moon"] = +1
-        if self.mass < 10:
-            self.surface = PlanetaryFeature(self)
-        elif self.irregular:
-            self.surface = Craters(self)
-        else:
-            self.surface = self.planet.vista.coords.choice(
-                [PlanetaryFeature, Clouds, Continents, Craters],
-                p=[0.1, 0.05, 0.15, 0.7],
-            )(self)
-        self.im = self.generate_sat()
-
-        # Setting the orbit's position and velocity.
-        self.orbit_velocity = 1  # self.planet.vista.coords.integers(1, 3)
-        self.orbit_radius = self.planet.mass * 2 / self.orbit_velocity
-        self.orbit_matrix = np.array(
-            [
-                [
-                    np.cos(np.radians(self.planet.tilt_from_x)),
-                    -1 * np.sin(np.radians(self.planet.tilt_from_x)),
-                ],
-                [
-                    np.sin(np.radians(self.planet.tilt_from_x)),
-                    np.cos(np.radians(self.planet.tilt_from_x)),
-                ],
-            ]
-        )
-        self.start_location = self.cluster.get_start_location(self)
-
-    def generate_sat(self):
-        """
-        A method called when creating the Satellite to generate the Satellite's image.
-        As this Satellite is a moon, it will have some randomly generated
-        PlanetaryFeatures.
-
-        Returns an PIL.Image of the Satellite with its static features.
-        """
-        sat_im = Image.new("RGBA", (self.mass + 10, self.mass + 10), (0, 0, 0, 0))
-        sat_mask = sat_im.copy()
-        draw_sat = ImageDraw.Draw(sat_im)
-        draw_mask = ImageDraw.Draw(sat_mask)
-        x_shift = self.irregular * self.vista.coords.integers(
-            self.mass // -5, self.mass // 5 + 1
-        )
-        y_shift = self.irregular * self.vista.coords.integers(
-            self.mass // -5, self.mass // 5 + 1
-        )
-        center = np.array((self.mass + 10, self.mass + 10)) // 2
-        self.surface.background(center, sat_im, draw_sat)
-        draw_sat.ellipse(
-            [5, 5, self.mass + x_shift, self.mass + y_shift],
-            fill=(*self.fill, 255),
-        )
-        draw_mask.ellipse(
-            [5, 5, self.mass + x_shift, self.mass + y_shift],
-            fill=(*self.fill, 255),
-        )
-
-        # A small proportion of moons will be irregular in shape.
-        if self.irregular:
-            print("Irregular moon!")
-            style = self.vista.coords.choice(
-                [
-                    (  # Cut-out Style
-                        [
-                            self.vista.coords.integers(
-                                self.mass // -10, self.mass // 10 + 1
-                            ),
-                            self.vista.coords.integers(
-                                self.mass // -10, self.mass // 10 + 1
-                            ),
-                            self.mass + 4 + x_shift // 3,
-                            self.mass + 4 + y_shift // 3,
-                        ],
-                        (0, 0, 0, 0),
-                    ),
-                    (  # Add-on style
-                        [
-                            self.vista.coords.integers(0, self.mass // 8 + 1),
-                            self.vista.coords.integers(0, self.mass // 8 + 1),
-                            self.mass + 4 + x_shift // 2,
-                            self.mass + 4 + y_shift // 2,
-                        ],
-                        (*self.fill, 255),
-                    ),
-                ]
-            ).tolist()
-            draw_mask.ellipse(style[0], tuple(style[1]))
-            if self.vista.coords.integers(2) == 0:
-                sat_im.transpose(Image.FLIP_TOP_BOTTOM)
-            sat_im.rotate(
-                self.vista.coords.integers(360),
-                resample=Image.BICUBIC,
-                expand=True,
-                center=(self.mass // 2 + 2, self.mass // 2 + 2),
-                fillcolor=(0, 0, 0, 0),
-            )
-        self.surface.foreground(center, sat_im, draw_sat, sat_mask)
-        sat_im = sat_im.rotate(
-            self.planet.tilt_from_x,
-            resample=Image.BICUBIC,
-            expand=True,
-            center=(self.mass // 2 + 2, self.mass // 2 + 2),
-            fillcolor=(0, 0, 0, 0),
-        )
-        return sat_im
-
-    def oribit_shift(self, x, y, image, frame_n, phase):
-        """
-        A method called once per frame to determine where the Satellite should be
-        draw on that frame.
-
-        Returns xy screen coordinates
-        """
-        # Find where the sat is in its orbit around the origin in Cartesian
-        shift_x = self.orbit_radius * np.cos(phase)
-        shift_y = (
-            self.orbit_radius
-            * abs(np.sin(np.radians(self.planet.tilt_from_y)))
-            * self.planet.spin
-            * np.sin(phase)
-        )
-        # Rotate around the planet's center to fit the planet's tilt from x
-        xy = [shift_x, shift_y] @ self.orbit_matrix
-        # Translate origin from 0,0 to the planet's top-right corner
-        xy += [x, y]
-        # Translate origin from planet's top-right corner to planet's center
-        xy += np.array(self.planet.im.size) // 2
-        # Offset by the distance from the sat's top-right to its center
-        xy -= np.array(self.im.size) // 2
-        return xy
-
-    def drifting_background(self, x, y, image, drawing_frame, frame_n):
-        """
-        Called every frame to draw Satellites that might be occuluded by the planet
-        as they move with respect to that planet.
-        """
-        phase = (
-            (self.orbit_velocity * frame_n + self.start_location)
-            / self.planet.vista.length
-            * 2
-            * np.pi
-        )
-        if (
-            self.planet.spin
-            * np.sin(np.radians(self.planet.tilt_from_y))
-            * np.sin(phase)
-            <= 0
-        ):
-            x, y = self.oribit_shift(x, y, image, frame_n, phase)
-            image.paste(self.im, (int(x), int(y)), mask=self.im)
-
-    def drifting_foreground(self, x, y, image, drawing_frame, frame_n):
-        """
-        Called every frame to draw Satellites that might pass in front of the planet
-        as they move with respect to that planet.
-        """
-        phase = (
-            (self.orbit_velocity * frame_n + self.start_location)
-            / self.planet.vista.length
-            * 2
-            * np.pi
-        )
-        if (
-            self.planet.spin
-            * np.sin(np.radians(self.planet.tilt_from_y))
-            * np.sin(phase)
-            > 0
-        ):
-            x, y = self.oribit_shift(x, y, image, frame_n, phase)
-            image.paste(self.im, (int(x), int(y)), mask=self.im)
-
-
-class Continents(PlanetaryFeature):
-    """
-    A static PlanetaryFeature that resembles coastlines and continents.
-    """
-
-    def __init__(self, planet):
-        super().__init__(planet)
-        if self.planet.colorful:
-            self.land_colors = self.planet.vista.palette.get_color(
-                shade=self.planet.shade
-            )
-        else:
-            self.land_colors = self.planet.vista.palette.get_color(hue=self.planet.hue)
-        self.land_colors = [
-            lc + (255,) for lc in self.land_colors if lc != self.planet.fill
-        ]
-        self.n_continents = self.planet.vista.coords.integers(3, 13)
-        if self.planet.is_planetoid:
-            self.planet.vista.status_dict[f"Planet {self.planet.layer}"][
-                "land masses"
-            ] = self.n_continents
-
-    def foreground(self, center, im, draw_im, planet_mask):
-        """
-        Called once to draw roughly self.n_continents number of continents across
-        the surface of the planet's image.
-        """
-        surface = coastlines(
-            self.planet.vista.coords,
-            im,
-            self.n_continents,
-            self.planet.fill,
-            self.land_colors,
-        )
-        im.paste(surface, (0, 0), mask=planet_mask)
-
-
-class Clouds(PlanetaryFeature):
-    """
-    A static PlanetaryFeature that approximates great bands of clouds like those found
-    on gas giants, as well as a few stable elliptical storms.
-    """
-
-    def __init__(self, planet):
-        super().__init__(planet)
-        self.n_clouds = self.planet.vista.coords.integers(
-            1, max(1, int(np.log(self.planet.mass))) + 2
-        )
-        if self.planet.is_planetoid:
-            self.planet.vista.status_dict[f"Planet {self.planet.layer}"][
-                "cloud bands"
-            ] = (self.n_clouds * 2 + 1)
-            self.planet.vista.status_dict[f"Planet {self.planet.layer}"][
-                "expected storms"
-            ] = int(np.log10(self.planet.width))
-        self.cloud_size = (
-            self.planet.mass,
-            self.planet.mass // max(1, (self.n_clouds - 2)),
-        )
-        if self.planet.colorful:
-            self.cloud_colors = self.planet.vista.palette.get_color(
-                shade=self.planet.shade
-            )
-        else:
-            self.cloud_colors = self.planet.vista.palette.get_color(hue=self.planet.hue)
-        self.cloud_colors = [
-            cc + (255,) for cc in self.cloud_colors if cc != self.planet.fill
-        ]
-
-    def foreground(self, center, im, draw_im, planet_mask):
-        """
-        Called once to draw roughly self.n_clouds number of cloud bands across
-        the surface of the planet's image.
-        """
-        surface = im.copy()
-        clouds = [
-            (
-                n,
-                nebula_like(
-                    self.planet.vista.coords,
-                    foreground_color=tuple(
-                        self.planet.vista.coords.choice(self.cloud_colors).tolist()
-                    ),
-                    box=self.cloud_size,
-                    background_color=self.planet.fill + (255,),
-                    ensure_wrap=False,
-                    eddies=True,
-                    eddy_colors=self.cloud_colors,
-                ),
-            )
-            for n in range(self.n_clouds)
-        ]
-        for n, cloud in clouds:
-            left_x, top_y = center - self.planet.mass // 2
-            surface.alpha_composite(
-                cloud, dest=(left_x, top_y + self.cloud_size[1] * n)
-            )
-        im.paste(surface, (0, 0), mask=planet_mask)
-
-
-class Craters(PlanetaryFeature):
-    """
-    A static PlanetaryFeature that resembles volcanic or impact craters.
-    """
-
-    def __init__(self, planet):
-        super().__init__(planet)
-        self.n_craters = self.planet.vista.coords.integers(1, self.planet.mass // 2 + 2)
-        if self.planet.is_planetoid:
-            self.planet.vista.status_dict[f"Planet {self.planet.layer}"][
-                "craters"
-            ] = self.n_craters
-
-    def foreground(self, center, im, draw_im, planet_mask):
-        """
-        Called once to draw roughly self.n_ccraters number of craters across
-        the surface of the planet's image.
-        """
-        surface = im.copy()
-        scar_surface = ImageDraw.Draw(surface)
-        left_x, top_y = center - self.planet.mass // 2
-        right_x, bottom_y = center + self.planet.mass // 2
-        for crater in range(self.n_craters):
-            diameter = int(
-                self.planet.vista.coords.normal(
-                    self.planet.mass / 8, self.planet.mass / 8
-                )
-            )
-            x = self.planet.vista.coords.integers(left_x - diameter, right_x)
-            y = self.planet.vista.coords.integers(top_y - diameter, bottom_y)
-            hue = self.planet.hue
-            shade = self.planet.vista.coords.choice(
-                [
-                    shade
-                    for shade in range(self.planet.vista.palette.shade_depth)
-                    if shade != self.planet.shade
-                ]
-            )
-            width = diameter // 10 + self.planet.vista.coords.integers(3)
-            outline = self.planet.vista.palette.get_color(hue=hue, shade=shade) + (255,)
-            scar_surface.ellipse(
-                [x, y, x + diameter, y + diameter],
-                fill=self.planet.fill,
-                outline=outline,
-                width=width,
-            )
-            ripples = [
-                self.planet.vista.coords.integers(0, 360, 2)
-                for trash in range(self.planet.vista.coords.integers(3))
-            ]
-            decay = 0
-            while width > decay:
-                decay += 1
-                for ripple in ripples:
-                    scar_surface.arc(
-                        [
-                            x - width,
-                            y - width,
-                            x + width + diameter,
-                            y + width + diameter,
-                        ],
-                        start=ripple[0] + 2 * decay,
-                        end=ripple[1] - 2 * decay,
-                        fill=outline,
-                        width=width - decay,
-                    )
-        im.paste(surface, mask=planet_mask)
-
-
-class Rings(PlanetaryFeature):
-    """
-    A static PlanetaryFeature that resembles rings around the planet.
-    """
-
-    def __init__(self, planet, med_diameter=None, local_height=None):
-        super().__init__(planet)
-        if self.planet.colorful:
-            hue = self.planet.vista.coords.choice(
-                [
-                    hue
-                    for hue in range(self.planet.vista.palette.hue_depth)
-                    if hue != self.planet.hue
-                ]
-            )
-            shade = self.planet.shade
-        else:
-            hue = self.planet.hue
-            shade = self.planet.vista.coords.choice(
-                [
-                    shade
-                    for shade in range(self.planet.vista.palette.shade_depth - 2)
-                    if shade != self.planet.shade
-                ]
-            )
-        self.fill = self.planet.vista.palette.get_color(hue=hue, shade=shade)
-        self.alpha = (255,)
-        self.thickness = self.planet.vista.coords.integers(1, self.planet.mass // 4 + 2)
-        if med_diameter == None:
-            self.med_diameter = self.planet.mass * 2
-        else:
-            self.med_diameter = med_diameter
-        self.diameter = (
-            int(
-                self.planet.vista.coords.normal(self.med_diameter, self.planet.mass / 6)
-            )
-            + self.thickness
-        )
-        if local_height == None:
-            self.ring_height = int(
-                self.diameter * np.sin(np.radians(self.planet.tilt_from_y))
-            )
-            if self.ring_height == 0:
-                self.ring_height = -1
-        else:
-            self.ring_height = local_height
-        self.sys_width = self.diameter + self.thickness * 2
-        self.sys_height = abs(self.ring_height) + self.thickness * 2
-        self.center = np.array((self.diameter, abs(self.ring_height))) // 2
-        self.planet.vista.status_dict[f"Planet {self.planet.layer}"]["rings"] += 1
-
-    def ring_maker(self, im, center):
-        """
-        The method used to draw the rings according to the specifications on the static
-        image of the planet.
-        """
-        x, y = center - self.center
-        self.ring = Image.new("RGBA", im.size, (0, 0, 0, 0))
-        self.draw_ring = ImageDraw.Draw(self.ring)
-        self.draw_ring.ellipse(
-            [x, y, x + self.diameter, y + abs(self.ring_height)],
-            fill=(0, 0, 0, 0),
-            outline=self.fill + self.alpha,
-            width=self.thickness,
-        )
-        # Cut-out for planet
-        clearance = abs(
-            int(self.planet.mass * np.sin(np.radians(self.planet.tilt_from_y)))
-        )
-        co_x = center[0] - self.planet.width // 2
-        co_y = center[1] - clearance // 2
-        self.draw_ring.ellipse(
-            [co_x - 1, co_y - 1, co_x + self.planet.mass + 2, co_y + clearance + 2],
-            fill=(0, 0, 0, 0),
-        )
-
-    def background(self, center, im, draw_im):
-        """
-        Called first to draw the full ring behind the planet.
-        """
-        self.ring_maker(im, center)
-        im.alpha_composite(self.ring)
-
-    def foreground(self, center, im, draw_im, planet_mask):
-        """
-        Called second to redraw the portion of the ring that should be in front
-        of the planet.
-        """
-        x, y = center - self.center
-        background = [0, 0, im.size[0], center[1]]
-        self.draw_ring.rectangle(background, fill=(0, 0, 0, 0))
-        if self.ring_height > 0:
-            self.ring.transpose(Image.FLIP_TOP_BOTTOM)
-        im.alpha_composite(self.ring)
-
-
 class Interior(CelestialBody):
     """
     The base class for an interior. Can be used for any interior that has no animated
@@ -1717,106 +1116,6 @@ class ComplementaryPalette16(Palette):
 
 
 ## Helper Functions
-
-
-def nebula_like(
-    coords,
-    foreground_color,
-    box=(240, 150),
-    background_color=(0, 0, 0, 0),
-    ensure_wrap=False,
-    eddies=False,
-    eddy_colors=None,
-):
-    sketch = Image.new("RGBA", box, (0, 0, 0, 0))
-    sketcher = ImageDraw.Draw(sketch)
-    width, height = box[0], box[1]
-    for level in (height // 3, height // 3 * 2):
-        attempts = 7
-        while attempts:
-            try:
-                attempts -= 1
-                xs = np.linspace(1, width, 8, endpoint=False)
-                # add caps to ensure the nebula behaves at the ends and wraps around
-                xs = (
-                    np.linspace(-width, 0, 8).tolist()
-                    + [x + coords.integers(-width // 24, width // 24) for x in xs]
-                    + np.linspace(width, width * 2, 8).tolist()
-                )
-                ys = (
-                    [level] * 8
-                    + [coords.normal(level, height // 36) for x in xs[8:-8]]
-                    + [level] * 8
-                )
-                poly = np.linalg.solve(np.vander(xs), ys).tolist()
-                break
-            except np.linalg.LinAlgError:
-                continue
-        else:
-            return sketch
-        points = [
-            (x, reduce(lambda a, b: x * (a + b), poly[:-1], 0) + poly[-1])
-            for x in range(width)
-        ]
-        sketcher.line(points, fill=foreground_color, joint="curve", width=2)
-    for x in range(1, width + 2, width // 2):
-        ImageDraw.floodfill(sketch, (x, int(height / 2)), foreground_color)
-    # Adding eddies and thousand-year-old storms
-    if eddies:
-        if eddies is True:
-            eddies = coords.integers(int(np.log10(width)) + 1)
-        for ed in range(eddies):
-            if eddy_colors is None:
-                eddy_color = foreground_color
-            else:
-                eddy_color = tuple(coords.choice(eddy_colors).tolist())
-            zone = width // eddies
-            ed_height = coords.integers(height // 6, height // 3)
-            ed_width = coords.integers(zone // 6, zone // 3)
-            if ed_height > ed_width:
-                ed_width, ed_height = ed_height, ed_width
-            x = coords.integers(zone * ed, max(2, zone * (1 + ed) - ed_width))
-            y = coords.integers(
-                height // 3 - ed_height // 2, height // 3 * 2 - ed_height // 2
-            )
-            sketcher.ellipse(
-                [x, y, x + ed_width, y + ed_height],
-                fill=eddy_color,
-                outline=background_color,
-                width=coords.integers(0, int(np.log10(width)) + 2),
-            )
-    # Let's smooth it out a bit
-    sketch = sketch.filter(ImageFilter.SMOOTH)
-    return sketch
-
-
-def coastlines(coords, im, n_continents, sea_color, land_colors):
-    sketch = Image.new("RGBA", im.size, sea_color)
-    sketcher = ImageDraw.Draw(sketch)
-    for n in range(n_continents):
-        x = coords.integers(0, im.size[0])
-        y = coords.integers(0, im.size[1])
-        path = []
-        for n in range(60 * im.size[0]):
-            new_x, new_y = x, y
-            while new_x == x and new_y == y:
-                new_x += coords.integers(-1, 2)
-                new_y += coords.integers(-1, 2)
-            x, y = new_x, new_y
-            path.append(x)
-            path.append(y)
-        if isinstance(land_colors, list):
-            land_color = tuple(coords.choice(land_colors).tolist())
-        else:
-            land_color = land_colors
-        sketcher.line(path, fill=land_color, width=2, joint="curve")
-    sketch = (
-        sketch.effect_spread(3)
-        # .filter(ImageFilter.GaussianBlur(3))
-        .filter(ImageFilter.SMOOTH)
-    )
-    return sketch
-
 
 def flame_like_alpha(full_size, box, left_curl=False, whisp=False):
     # Make some noise!
